+++ src/GoldyBot/utility/views/forms.py
@@ -28,7 +28,7 @@ def normal_form(title:str, items:List[GoldyBot.nextcord.ui.Item], callback:Calla
 
             if view.value == True:
                 if not callback == None:
-                    await callback(self.values) # Execute callback
+                    await callback((lambda x: [item_.value for item_ in x])(items)) # Execute callback
 
                     GoldyBot.logging.log(f"[{MODULE_NAME}] Normal form modal for '{author.name}' executed it's function '{callback.__name__}'!")