+++ src/CHANGELOG.rst
@@ -9,6 +9,7 @@ v4.0dev25 = New Dev Release (???)
 - Added dropdown view function in goldy bot utils.
 - Improved GoldyBot dropdown view.
 - Added LIME_GREEN.
+- Added GoldyBot "send_modal()" method in commands utils.
 
 v4.0dev24 = New Dev Release (01/10/2022)
 ---------------------------